requirements.txt
fastapi
uvicorn
requests
python-dotenv
retell-sdk
flask
